% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_plate_data.R
\docType{data}
\name{example_plate_data}
\alias{example_plate_data}
\title{Example Plate Data}
\format{
A data frame with X rows and 3 variables:
\describe{
\item{SampleID}{Unique identifier for each sample}
\item{PlateID}{Identifier for the plate}
\item{WellID}{Well position on the plate}
}
}
\source{
{From the authors of the package}
}
\usage{
example_plate_data
}
\description{
A dataset containing unprocessed plate data for multiple source plates. EMTPY wells have a SampleID of "EMPTY"
}
\keyword{datasets}
