% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result_from.R
\name{result_from}
\alias{result_from}
\title{Take a Hamilton input file and return the resulting destination plate.}
\usage{
result_from(data, platename = NULL)
}
\arguments{
\item{data}{A tibble or file path to a Hamilton input file. Must contain columns SampleID,
PlateID, SourceWellID, and DestWellID.}

\item{platename}{A character string with the name of the destination plate being created.
If not provided and data is a file path, the filename will be used as the plate name.
If not provided and data is a tibble, "NewPlate" will be used as the default name.}
}
\value{
Returns a tibble with columns SampleID, PlateID, and WellID.
}
\description{
This function takes a Hamilton robot input file (or equivalent tibble) and creates a
representation of the destination plate after the robot has performed the cherrypicking
operation. It is useful for tracking samples through the laboratory workflow and
understanding the resulting plate layout.
}
\examples{
\dontrun{
data <- tibble::tibble(
  SampleID = c("C234323CVP", "C240026CVP", "C240036CVP", "C240199CVP"),
  PlateID = c("CVP20240204CDFW01", "CVP20240204CDFW01", "CVP20240204CDFW01", "CVP20240218CDFW02"),
  SourceWellID = c("A1", "A2", "A3","A1"),
  DestWellID = c("A1", "A2", "A3","A4")
)
result_plate <- result_from(data)
result_plate
}
}
