% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_hamilton_csvs.R
\name{write_hamilton_csvs}
\alias{write_hamilton_csvs}
\title{Write a tibble to multiple CSV files for Hamilton robot input}
\usage{
write_hamilton_csvs(
  data,
  rows_per_file = 93,
  max_plate_ids = 20,
  file_prefix = "plate_input_",
  auto_dest_well = TRUE
)
}
\arguments{
\item{data}{A tibble to be split and written to CSV files.}

\item{rows_per_file}{Maximum number of rows per file. Default is 93.}

\item{max_plate_ids}{Maximum number of unique plate IDs per file. Default is 20.}

\item{file_prefix}{Prefix for the output CSV filenames. Default is "plate_input_".}

\item{auto_dest_well}{If TRUE, the destination well will be automatically filled in sequentially, i.e. A1,B1,C1, etc. Default is TRUE. If FALSE, then a blank column will be created for the destination well which must be filled in by the user afterwards.}
}
\value{
Nothing is returned, but CSV files are written to the current working directory.
}
\description{
This function splits a tibble into multiple CSV files based on specified criteria,
formatting the output for use with Hamilton robots.
}
\examples{
\dontrun{
data <- tibble::tibble(
  SampleID = c("C234323CVP", "C240026CVP", "C240036CVP", "C240199CVP"),
  PlateID = c("CVP20240204CDFW01", "CVP20240204CDFW01", "CVP20240204CDFW01", "CVP20240218CDFW02"),
  WellID = c("A1", "A2", "A3","A1")
)
write_hamilton_csvs(data)
}
}
